requirements.txt
# flask ecosystem
flask
flask-crontab
flask-restful
flasgger

# net
requests

# env
python-dotenv